requirements.txt
joblib
streamlit
pandas
scikit-learn
matplotlib
numpy
seaborn
gspread
google-auth
